+++ app/src/main/java/com/aconno/acnsensa/dagger/gcloudpublisher/GoogleCloudPublisherComponent.kt
@@ -1,7 +1,7 @@
 package com.aconno.acnsensa.dagger.gcloudpublisher
 
 import com.aconno.acnsensa.dagger.application.AppComponent
-import com.aconno.acnsensa.ui.settings.selectpublish.GoogleCloudPublisherActivity
+import com.aconno.acnsensa.ui.settings.publishers.selectpublish.GoogleCloudPublisherActivity
 import dagger.Component
 
 @Component(dependencies = [AppComponent::class], modules = [GoogleCloudPublisherModule::class])
