+++ app/src/main/java/com/aconno/acnsensa/ui/settings/publishers/PublishListActivity.kt
@@ -1,21 +1,18 @@
-package com.aconno.acnsensa.ui.settings
+package com.aconno.acnsensa.ui.settings.publishers
 
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
 import android.support.v4.app.Fragment
 import android.support.v7.app.AppCompatActivity
-import android.view.Menu
-import android.view.MenuItem
 import com.aconno.acnsensa.R
 import com.aconno.acnsensa.model.BasePublishModel
 import com.aconno.acnsensa.model.GooglePublishModel
 import com.aconno.acnsensa.model.MqttPublishModel
 import com.aconno.acnsensa.model.RESTPublishModel
-import com.aconno.acnsensa.ui.settings.selectpublish.GoogleCloudPublisherActivity
-import com.aconno.acnsensa.ui.settings.selectpublish.MqttPublisherActivity
-import com.aconno.acnsensa.ui.settings.selectpublish.RESTPublisherActivity
-import com.aconno.acnsensa.ui.settings.selectpublish.SelectPublisherActivity
+import com.aconno.acnsensa.ui.settings.publishers.selectpublish.GoogleCloudPublisherActivity
+import com.aconno.acnsensa.ui.settings.publishers.selectpublish.MqttPublisherActivity
+import com.aconno.acnsensa.ui.settings.publishers.selectpublish.RESTPublisherActivity
 import kotlinx.android.synthetic.main.activity_add_publish.*
 
 
@@ -34,7 +31,8 @@ class PublishListActivity : AppCompatActivity(),
         if (fm != null) {
             var fragment: Fragment? = fm.findFragmentById(R.id.publish_list_container)
             if (fragment == null) {
-                fragment = PublishListFragment.newInstance()
+                fragment =
+                        PublishListFragment.newInstance()
                 fm.beginTransaction().add(R.id.publish_list_container, fragment).commit()
             }
         }
