+++ app/src/main/java/com/aconno/acnsensa/ui/settings/publishers/rheader/ItemClickListenerWithPos.kt
@@ -1,4 +1,4 @@
-package com.aconno.acnsensa.ui.settings.rheader
+package com.aconno.acnsensa.ui.settings.publishers.rheader
 
 interface ItemClickListenerWithPos<in T> {
     fun onItemClick(position: Int, item: T?)
