+++ app/src/main/java/com/aconno/acnsensa/ui/settings/publishers/selectpublish/GoogleCloudPublisherActivity.kt
@@ -1,48 +1,51 @@
-package com.aconno.acnsensa.ui.settings.selectpublish
+package com.aconno.acnsensa.ui.settings.publishers.selectpublish
 
+import android.app.Activity
 import android.app.AlertDialog
 import android.content.Context
 import android.content.Intent
+import android.net.Uri
 import android.os.Bundle
 import android.view.Menu
 import android.view.MenuItem
 import android.view.View
-import android.widget.RadioButton
 import android.widget.Switch
 import android.widget.TextView
 import android.widget.Toast
 import com.aconno.acnsensa.AcnSensaApplication
 import com.aconno.acnsensa.R
-import com.aconno.acnsensa.dagger.mqttpublisher.DaggerMqttPublisherComponent
-import com.aconno.acnsensa.dagger.mqttpublisher.MqttPublisherComponent
-import com.aconno.acnsensa.dagger.mqttpublisher.MqttPublisherModule
+import com.aconno.acnsensa.dagger.gcloudpublisher.DaggerGoogleCloudPublisherComponent
+import com.aconno.acnsensa.dagger.gcloudpublisher.GoogleCloudPublisherComponent
+import com.aconno.acnsensa.dagger.gcloudpublisher.GoogleCloudPublisherModule
 import com.aconno.acnsensa.data.converter.PublisherIntervalConverter
-import com.aconno.acnsensa.data.publisher.MqttPublisher
+import com.aconno.acnsensa.data.publisher.GoogleCloudPublisher
 import com.aconno.acnsensa.domain.Publisher
 import com.aconno.acnsensa.domain.model.Device
 import com.aconno.acnsensa.model.DeviceRelationModel
-import com.aconno.acnsensa.model.MqttPublishModel
-import com.aconno.acnsensa.model.mapper.MqttPublishModelDataMapper
+import com.aconno.acnsensa.model.GooglePublishModel
+import com.aconno.acnsensa.model.mapper.GooglePublishModelDataMapper
 import com.aconno.acnsensa.ui.base.BaseActivity
-import com.aconno.acnsensa.viewmodel.MqttPublisherViewModel
+import com.aconno.acnsensa.viewmodel.GoogleCloudPublisherViewModel
 import io.reactivex.Completable
 import io.reactivex.CompletableObserver
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.disposables.Disposable
 import io.reactivex.functions.Consumer
 import io.reactivex.schedulers.Schedulers
-import kotlinx.android.synthetic.main.activity_mqtt_publisher.*
-import kotlinx.android.synthetic.main.layout_mqtt.*
+import kotlinx.android.synthetic.main.activity_google_cloud_publisher.*
+import kotlinx.android.synthetic.main.layout_google.*
+import java.security.KeyFactory
+import java.security.spec.PKCS8EncodedKeySpec
 import java.text.SimpleDateFormat
 import java.util.*
 import javax.inject.Inject
 
-class MqttPublisherActivity : BaseActivity() {
+class GoogleCloudPublisherActivity : BaseActivity() {
 
     @Inject
-    lateinit var mqttPublisherViewModel: MqttPublisherViewModel
+    lateinit var googleViewModel: GoogleCloudPublisherViewModel
 
-    private var mqttPublishModel: MqttPublishModel? = null
+    private var googlePublishModel: GooglePublishModel? = null
     private lateinit var deviceList: List<DeviceRelationModel>
     private var isTestingAlreadyRunning: Boolean = false
 
@@ -51,7 +54,7 @@ class MqttPublisherActivity : BaseActivity() {
             progressbar.visibility = View.INVISIBLE
             isTestingAlreadyRunning = false
             Toast.makeText(
-                this@MqttPublisherActivity,
+                this@GoogleCloudPublisherActivity,
                 getString(R.string.test_succeeded),
                 Toast.LENGTH_SHORT
             ).show()
@@ -61,7 +64,7 @@ class MqttPublisherActivity : BaseActivity() {
             progressbar.visibility = View.INVISIBLE
             isTestingAlreadyRunning = false
             Toast.makeText(
-                this@MqttPublisherActivity,
+                this@GoogleCloudPublisherActivity,
                 getString(R.string.test_failed),
                 Toast.LENGTH_SHORT
             ).show()
@@ -72,29 +75,29 @@ class MqttPublisherActivity : BaseActivity() {
         }
     }
 
-    private val mqttPublisherComponent: MqttPublisherComponent by lazy {
+    private val googleCloudPublisherComponent: GoogleCloudPublisherComponent by lazy {
         val acnSensaApplication: AcnSensaApplication? = application as? AcnSensaApplication
 
-        DaggerMqttPublisherComponent
+        DaggerGoogleCloudPublisherComponent
             .builder()
             .appComponent(acnSensaApplication?.appComponent)
-            .mqttPublisherModule(MqttPublisherModule(this))
+            .googleCloudPublisherModule(GoogleCloudPublisherModule(this))
             .build()
     }
 
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_mqtt_publisher)
-        mqttPublisherComponent.inject(this)
+        setContentView(R.layout.activity_google_cloud_publisher)
+        googleCloudPublisherComponent.inject(this)
 
         setSupportActionBar(custom_toolbar)
         supportActionBar!!.setDisplayShowTitleEnabled(true)
 
         initViews()
-        if (intent.hasExtra(MqttPublisherActivity.MQTT_PUBLISHER_ACTIVITY_KEY)) {
-            mqttPublishModel =
-                    intent.getParcelableExtra(MqttPublisherActivity.MQTT_PUBLISHER_ACTIVITY_KEY)
+        if (intent.hasExtra(GOOGLE_PUBLISHER_ACTIVITY_KEY)) {
+            googlePublishModel =
+                    intent.getParcelableExtra(GOOGLE_PUBLISHER_ACTIVITY_KEY)
+
             setFields()
         }
     }
@@ -105,7 +108,7 @@ class MqttPublisherActivity : BaseActivity() {
 
         if (menu != null) {
             val item = menu.findItem(R.id.action_publish_done)
-            if (mqttPublishModel != null) {
+            if (googlePublishModel != null) {
                 item.title = getString(R.string.update)
             }
         }
@@ -124,7 +127,45 @@ class MqttPublisherActivity : BaseActivity() {
         return super.onOptionsItemSelected(item)
     }
 
+    /**
+     * This method is called after @Intent.ACTION_OPEN_DOCUMENT result is returned.
+     */
+    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
+        if (resultCode == Activity.RESULT_OK && requestCode == PICKFILE_REQUEST_CODE) {
+            data?.let {
+                val path = it.data.toString()
+
+                applicationContext.contentResolver.takePersistableUriPermission(
+                    data.data,
+                    Intent.FLAG_GRANT_READ_URI_PERMISSION or Intent.FLAG_GRANT_WRITE_URI_PERMISSION
+                )
+
+                if (isFileValidPKCS8(getPrivateKeyData(it.data.toString()))) {
+                    edit_privatekey.text = path
+                } else {
+                    Toast.makeText(
+                        this,
+                        getString(R.string.not_valid_file_pkcs8),
+                        Toast.LENGTH_SHORT
+                    ).show()
+                }
+            }
+        }
+        super.onActivityResult(requestCode, resultCode, data)
+    }
+
     private fun initViews() {
+        edit_privatekey.setOnClickListener {
+
+            val intent = Intent(Intent.ACTION_OPEN_DOCUMENT)
+            intent.type = "*/*"
+            startActivityForResult(
+                intent,
+                PICKFILE_REQUEST_CODE
+            )
+
+        }
+
         btn_info.setOnClickListener {
             val builder = AlertDialog.Builder(this)
 
@@ -139,18 +180,16 @@ class MqttPublisherActivity : BaseActivity() {
         }
 
         //Load Devices
-        val subscribe = mqttPublisherViewModel.getAllDevices()
+        val subscribe = googleViewModel.getAllDevices()
             .subscribeOn(Schedulers.io())
             .observeOn(AndroidSchedulers.mainThread())
             .subscribe(Consumer {
                 deviceList = it!!
                 addDevices(deviceList)
 
-                if (mqttPublishModel != null) {
+                if (googlePublishModel != null) {
                     addDisposable(
-                        mqttPublisherViewModel.getDevicesThatConnectedWithMqttPublish(
-                            mqttPublishModel!!.id
-                        )
+                        googleViewModel.getDevicesThatConnectedWithPublish(googlePublishModel!!)
                             .subscribeOn(Schedulers.io())
                             .observeOn(AndroidSchedulers.mainThread())
                             .subscribe(Consumer {
@@ -205,47 +244,41 @@ class MqttPublisherActivity : BaseActivity() {
             .isChecked = true
     }
 
+
     private fun setFields() {
-        edit_name.setText(mqttPublishModel?.name)
+        edit_name.setText(googlePublishModel?.name)
 
         edit_interval_count.setText(
             PublisherIntervalConverter.calculateCountFromMillis(
-                mqttPublishModel!!.timeMillis,
-                mqttPublishModel!!.timeType
+                googlePublishModel!!.timeMillis,
+                googlePublishModel!!.timeType
             )
         )
 
         spinner_interval_time.setSelection(
             resources.getStringArray(R.array.PublishIntervals).indexOf(
-                mqttPublishModel?.timeType
+                googlePublishModel?.timeType
             )
         )
 
-        if (mqttPublishModel!!.lastTimeMillis == 0L) {
+        if (googlePublishModel!!.lastTimeMillis == 0L) {
             text_lastdatasent.visibility = View.GONE
         } else {
             text_lastdatasent.visibility = View.VISIBLE
             val str = getString(R.string.last_data_sent) + " " +
                     millisToFormattedDateString(
-                        mqttPublishModel!!.lastTimeMillis
+                        googlePublishModel!!.lastTimeMillis
                     )
             text_lastdatasent.text = str
         }
 
+        edit_projectid.setText(googlePublishModel!!.projectId)
+        edit_region.setText(googlePublishModel!!.region)
+        edit_deviceregistry.setText(googlePublishModel!!.deviceRegistry)
+        edit_device.setText(googlePublishModel!!.device)
+        edit_privatekey.text = googlePublishModel!!.privateKey
 
-        edit_url_mqtt.setText(mqttPublishModel!!.url)
-        edit_clientid_mqtt.setText(mqttPublishModel!!.clientId)
-        edit_username_mqtt.setText(mqttPublishModel!!.username)
-        edit_password_mqtt.setText(mqttPublishModel!!.password)
-        edit_topic_mqtt.setText(mqttPublishModel!!.topic)
-
-        when (mqttPublishModel!!.qos) {
-            0 -> qos_0.isChecked = true
-            1 -> qos_1.isChecked = true
-            2 -> qos_2.isChecked = true
-        }
-
-        edit_datastring.setText(mqttPublishModel?.dataString)
+        edit_datastring.setText(googlePublishModel?.dataString)
     }
 
     private fun millisToFormattedDateString(millis: Long): String {
@@ -256,11 +289,11 @@ class MqttPublisherActivity : BaseActivity() {
     }
 
     private fun addOrUpdate() {
-        val mqttPublishModel = toMqttPublishModel()
-        if (mqttPublishModel != null) {
-            mqttPublisherViewModel.save(mqttPublishModel)
+        val googlePublishModel = toGooglePublishModel()
+        if (googlePublishModel != null) {
+            googleViewModel.save(googlePublishModel)
                 .flatMapCompletable {
-                    addRelationsToMqtt(it)
+                    addRelationsToGoogle(it)
                 }.subscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread())
                 .subscribe(object : CompletableObserver {
@@ -277,7 +310,7 @@ class MqttPublisherActivity : BaseActivity() {
                     override fun onError(e: Throwable) {
                         progressbar.visibility = View.INVISIBLE
                         Toast.makeText(
-                            this@MqttPublisherActivity,
+                            this@GoogleCloudPublisherActivity,
                             e.message,
                             Toast.LENGTH_SHORT
                         )
@@ -287,58 +320,24 @@ class MqttPublisherActivity : BaseActivity() {
         }
     }
 
-    private fun addRelationsToMqtt(it: Long): Completable {
-        val count = layout_devices.childCount
-
-        val setOfCompletable = mutableSetOf<Completable>()
-
-        for (i in 0..(count - 1)) {
-            val deviceRelationModel = deviceList[i]
-
-            val isChecked =
-                layout_devices.getChildAt(i).findViewById<Switch>(R.id.switch_device).isChecked
-
-            if (isChecked) {
-                setOfCompletable.add(
-                    mqttPublisherViewModel.addOrUpdateMqttRelation(
-                        deviceId = deviceRelationModel.macAddress,
-                        mqttId = it
-                    )
-                )
-
-            } else {
-                setOfCompletable.add(
-                    mqttPublisherViewModel.deleteRelationMqtt(
-                        deviceId = deviceRelationModel.macAddress,
-                        mqttId = it
-                    )
-                )
-            }
-        }
-
-        return Completable.merge(setOfCompletable)
-    }
-
-    private fun toMqttPublishModel(): MqttPublishModel? {
+    private fun toGooglePublishModel(): GooglePublishModel? {
         val name = edit_name.text.toString().trim()
-        val url = edit_url_mqtt.text.toString().trim()
-        val clientId = edit_clientid_mqtt.text.toString().trim()
-        val username = edit_username_mqtt.text.toString().trim()
-        val password = edit_password_mqtt.text.toString().trim()
-        val topic = edit_topic_mqtt.text.toString().trim()
-        val qos =
-            findViewById<RadioButton>(radio_group_mqtt.checkedRadioButtonId).text.toString().toInt()
+        val projectId = edit_projectid.text.toString().trim()
+        val region = edit_region.text.toString().trim()
+        val deviceRegistry = edit_deviceregistry.text.toString().trim()
+        val device = edit_device.text.toString().trim()
+        val privateKey = edit_privatekey.text.toString().trim()
         val timeType = spinner_interval_time.selectedItem.toString()
         val timeCount = edit_interval_count.text.toString()
         val datastring = edit_datastring.text.toString()
 
-        if (mqttPublisherViewModel.checkFieldsAreEmpty(
+        if (googleViewModel.checkFieldsAreEmpty(
                 name,
-                url,
-                clientId,
-                username,
-                password,
-                topic,
+                projectId,
+                region,
+                deviceRegistry,
+                device,
+                privateKey,
                 timeType,
                 timeCount,
                 datastring
@@ -348,18 +347,18 @@ class MqttPublisherActivity : BaseActivity() {
             return null
         }
 
-        val id = if (mqttPublishModel == null) 0 else mqttPublishModel!!.id
+        val id = if (googlePublishModel == null) 0 else googlePublishModel!!.id
         val timeMillis = PublisherIntervalConverter.calculateMillis(timeCount, timeType)
-        val lastTimeMillis = if (mqttPublishModel == null) 0 else mqttPublishModel!!.lastTimeMillis
-        return MqttPublishModel(
+        val lastTimeMillis =
+            if (googlePublishModel == null) 0 else googlePublishModel!!.lastTimeMillis
+        return GooglePublishModel(
             id,
             name,
-            url,
-            clientId,
-            username,
-            password,
-            topic,
-            qos,
+            projectId,
+            region,
+            deviceRegistry,
+            device,
+            privateKey,
             false,
             timeType,
             timeMillis,
@@ -368,28 +367,78 @@ class MqttPublisherActivity : BaseActivity() {
         )
     }
 
+    private fun addRelationsToGoogle(gId: Long): Completable {
+        val count = layout_devices.childCount
+
+        val setOfCompletable: MutableSet<Completable> = mutableSetOf()
+
+        for (i in 0..(count - 1)) {
+            val deviceRelationModel = deviceList[i]
+
+            val isChecked =
+                layout_devices.getChildAt(i).findViewById<Switch>(R.id.switch_device).isChecked
+
+            if (isChecked) {
+                setOfCompletable.add(
+                    googleViewModel.addOrUpdateGoogleRelation(
+                        deviceId = deviceRelationModel.macAddress,
+                        googleId = gId
+                    )
+                )
+            } else {
+                setOfCompletable.add(
+                    googleViewModel.deleteRelationGoogle(
+                        deviceId = deviceRelationModel.macAddress,
+                        googleId = gId
+                    )
+                )
+            }
+        }
+        return Completable.merge(setOfCompletable)
+    }
+
+    private fun isFileValidPKCS8(byteArray: ByteArray): Boolean {
+        val spec = PKCS8EncodedKeySpec(byteArray)
+        val kf = KeyFactory.getInstance("RSA")
+
+        try {
+            kf.generatePrivate(spec)
+        } catch (e: Exception) {
+            return false
+        }
+        return true
+    }
+
+    private fun getPrivateKeyData(privateKey: String): ByteArray {
+        val uri = Uri.parse(privateKey)
+        val stream = contentResolver.openInputStream(uri)
+
+        val size = stream.available()
+        val buffer = ByteArray(size)
+        stream.read(buffer)
+        stream.close()
+        return buffer
+    }
+
     private fun test() {
         if (!isTestingAlreadyRunning) {
             isTestingAlreadyRunning = true
-
             Toast.makeText(this, getString(R.string.testings_started), Toast.LENGTH_SHORT).show()
+            val toGooglePublishModel = toGooglePublishModel()
 
-            val toMqttPublishModel = toMqttPublishModel()
-
-            if (toMqttPublishModel == null) {
+            if (toGooglePublishModel == null) {
                 isTestingAlreadyRunning = false
                 return
             }
 
-            testMqttConnection(toMqttPublishModel)
+            testGoogleConnection(toGooglePublishModel)
         }
-
     }
 
-    private fun testMqttConnection(toMqttPublishModel: MqttPublishModel) {
-        val publisher = MqttPublisher(
+    private fun testGoogleConnection(toGooglePublishModel: GooglePublishModel) {
+        val publisher = GoogleCloudPublisher(
             applicationContext,
-            MqttPublishModelDataMapper().toMqttPublish(toMqttPublishModel),
+            GooglePublishModelDataMapper().transform(toGooglePublishModel),
             listOf(Device("TestDevice", "Mac"))
         )
 
@@ -397,18 +446,18 @@ class MqttPublisherActivity : BaseActivity() {
         publisher.test(testConnectionCallback)
     }
 
-
     companion object {
         //This is used for the file selector intent
-        private const val MQTT_PUBLISHER_ACTIVITY_KEY = "MQTT_PUBLISHER_ACTIVITY_KEY"
+        const val PICKFILE_REQUEST_CODE: Int = 10213
+        private const val GOOGLE_PUBLISHER_ACTIVITY_KEY = "GOOGLE_PUBLISHER_ACTIVITY_KEY"
 
-        fun start(context: Context, mqttPublishModel: MqttPublishModel? = null) {
-            val intent = Intent(context, MqttPublisherActivity::class.java)
+        fun start(context: Context, googlePublishModel: GooglePublishModel? = null) {
+            val intent = Intent(context, GoogleCloudPublisherActivity::class.java)
 
-            mqttPublishModel?.let {
+            googlePublishModel?.let {
                 intent.putExtra(
-                    MQTT_PUBLISHER_ACTIVITY_KEY,
-                    mqttPublishModel
+                    GOOGLE_PUBLISHER_ACTIVITY_KEY,
+                    googlePublishModel
                 )
             }
