+++ app/src/main/java/com/aconno/acnsensa/ui/settings/publishers/selectpublish/MqttPublisherActivity.kt
@@ -1,51 +1,48 @@
-package com.aconno.acnsensa.ui.settings.selectpublish
+package com.aconno.acnsensa.ui.settings.publishers.selectpublish
 
-import android.app.Activity
 import android.app.AlertDialog
 import android.content.Context
 import android.content.Intent
-import android.net.Uri
 import android.os.Bundle
 import android.view.Menu
 import android.view.MenuItem
 import android.view.View
+import android.widget.RadioButton
 import android.widget.Switch
 import android.widget.TextView
 import android.widget.Toast
 import com.aconno.acnsensa.AcnSensaApplication
 import com.aconno.acnsensa.R
-import com.aconno.acnsensa.dagger.gcloudpublisher.DaggerGoogleCloudPublisherComponent
-import com.aconno.acnsensa.dagger.gcloudpublisher.GoogleCloudPublisherComponent
-import com.aconno.acnsensa.dagger.gcloudpublisher.GoogleCloudPublisherModule
+import com.aconno.acnsensa.dagger.mqttpublisher.DaggerMqttPublisherComponent
+import com.aconno.acnsensa.dagger.mqttpublisher.MqttPublisherComponent
+import com.aconno.acnsensa.dagger.mqttpublisher.MqttPublisherModule
 import com.aconno.acnsensa.data.converter.PublisherIntervalConverter
-import com.aconno.acnsensa.data.publisher.GoogleCloudPublisher
+import com.aconno.acnsensa.data.publisher.MqttPublisher
 import com.aconno.acnsensa.domain.Publisher
 import com.aconno.acnsensa.domain.model.Device
 import com.aconno.acnsensa.model.DeviceRelationModel
-import com.aconno.acnsensa.model.GooglePublishModel
-import com.aconno.acnsensa.model.mapper.GooglePublishModelDataMapper
+import com.aconno.acnsensa.model.MqttPublishModel
+import com.aconno.acnsensa.model.mapper.MqttPublishModelDataMapper
 import com.aconno.acnsensa.ui.base.BaseActivity
-import com.aconno.acnsensa.viewmodel.GoogleCloudPublisherViewModel
+import com.aconno.acnsensa.viewmodel.MqttPublisherViewModel
 import io.reactivex.Completable
 import io.reactivex.CompletableObserver
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.disposables.Disposable
 import io.reactivex.functions.Consumer
 import io.reactivex.schedulers.Schedulers
-import kotlinx.android.synthetic.main.activity_google_cloud_publisher.*
-import kotlinx.android.synthetic.main.layout_google.*
-import java.security.KeyFactory
-import java.security.spec.PKCS8EncodedKeySpec
+import kotlinx.android.synthetic.main.activity_mqtt_publisher.*
+import kotlinx.android.synthetic.main.layout_mqtt.*
 import java.text.SimpleDateFormat
 import java.util.*
 import javax.inject.Inject
 
-class GoogleCloudPublisherActivity : BaseActivity() {
+class MqttPublisherActivity : BaseActivity() {
 
     @Inject
-    lateinit var googleViewModel: GoogleCloudPublisherViewModel
+    lateinit var mqttPublisherViewModel: MqttPublisherViewModel
 
-    private var googlePublishModel: GooglePublishModel? = null
+    private var mqttPublishModel: MqttPublishModel? = null
     private lateinit var deviceList: List<DeviceRelationModel>
     private var isTestingAlreadyRunning: Boolean = false
 
@@ -54,7 +51,7 @@ class GoogleCloudPublisherActivity : BaseActivity() {
             progressbar.visibility = View.INVISIBLE
             isTestingAlreadyRunning = false
             Toast.makeText(
-                this@GoogleCloudPublisherActivity,
+                this@MqttPublisherActivity,
                 getString(R.string.test_succeeded),
                 Toast.LENGTH_SHORT
             ).show()
@@ -64,7 +61,7 @@ class GoogleCloudPublisherActivity : BaseActivity() {
             progressbar.visibility = View.INVISIBLE
             isTestingAlreadyRunning = false
             Toast.makeText(
-                this@GoogleCloudPublisherActivity,
+                this@MqttPublisherActivity,
                 getString(R.string.test_failed),
                 Toast.LENGTH_SHORT
             ).show()
@@ -75,29 +72,29 @@ class GoogleCloudPublisherActivity : BaseActivity() {
         }
     }
 
-    private val googleCloudPublisherComponent: GoogleCloudPublisherComponent by lazy {
+    private val mqttPublisherComponent: MqttPublisherComponent by lazy {
         val acnSensaApplication: AcnSensaApplication? = application as? AcnSensaApplication
 
-        DaggerGoogleCloudPublisherComponent
+        DaggerMqttPublisherComponent
             .builder()
             .appComponent(acnSensaApplication?.appComponent)
-            .googleCloudPublisherModule(GoogleCloudPublisherModule(this))
+            .mqttPublisherModule(MqttPublisherModule(this))
             .build()
     }
 
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_google_cloud_publisher)
-        googleCloudPublisherComponent.inject(this)
+        setContentView(R.layout.activity_mqtt_publisher)
+        mqttPublisherComponent.inject(this)
 
         setSupportActionBar(custom_toolbar)
         supportActionBar!!.setDisplayShowTitleEnabled(true)
 
         initViews()
-        if (intent.hasExtra(GOOGLE_PUBLISHER_ACTIVITY_KEY)) {
-            googlePublishModel =
-                    intent.getParcelableExtra(GOOGLE_PUBLISHER_ACTIVITY_KEY)
-
+        if (intent.hasExtra(MQTT_PUBLISHER_ACTIVITY_KEY)) {
+            mqttPublishModel =
+                    intent.getParcelableExtra(MQTT_PUBLISHER_ACTIVITY_KEY)
             setFields()
         }
     }
@@ -108,7 +105,7 @@ class GoogleCloudPublisherActivity : BaseActivity() {
 
         if (menu != null) {
             val item = menu.findItem(R.id.action_publish_done)
-            if (googlePublishModel != null) {
+            if (mqttPublishModel != null) {
                 item.title = getString(R.string.update)
             }
         }
@@ -127,45 +124,7 @@ class GoogleCloudPublisherActivity : BaseActivity() {
         return super.onOptionsItemSelected(item)
     }
 
-    /**
-     * This method is called after @Intent.ACTION_OPEN_DOCUMENT result is returned.
-     */
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        if (resultCode == Activity.RESULT_OK && requestCode == GoogleCloudPublisherActivity.PICKFILE_REQUEST_CODE) {
-            data?.let {
-                val path = it.data.toString()
-
-                applicationContext.contentResolver.takePersistableUriPermission(
-                    data.data,
-                    Intent.FLAG_GRANT_READ_URI_PERMISSION or Intent.FLAG_GRANT_WRITE_URI_PERMISSION
-                )
-
-                if (isFileValidPKCS8(getPrivateKeyData(it.data.toString()))) {
-                    edit_privatekey.text = path
-                } else {
-                    Toast.makeText(
-                        this,
-                        getString(R.string.not_valid_file_pkcs8),
-                        Toast.LENGTH_SHORT
-                    ).show()
-                }
-            }
-        }
-        super.onActivityResult(requestCode, resultCode, data)
-    }
-
     private fun initViews() {
-        edit_privatekey.setOnClickListener {
-
-            val intent = Intent(Intent.ACTION_OPEN_DOCUMENT)
-            intent.type = "*/*"
-            startActivityForResult(
-                intent,
-                GoogleCloudPublisherActivity.PICKFILE_REQUEST_CODE
-            )
-
-        }
-
         btn_info.setOnClickListener {
             val builder = AlertDialog.Builder(this)
 
@@ -180,16 +139,18 @@ class GoogleCloudPublisherActivity : BaseActivity() {
         }
 
         //Load Devices
-        val subscribe = googleViewModel.getAllDevices()
+        val subscribe = mqttPublisherViewModel.getAllDevices()
             .subscribeOn(Schedulers.io())
             .observeOn(AndroidSchedulers.mainThread())
             .subscribe(Consumer {
                 deviceList = it!!
                 addDevices(deviceList)
 
-                if (googlePublishModel != null) {
+                if (mqttPublishModel != null) {
                     addDisposable(
-                        googleViewModel.getDevicesThatConnectedWithPublish(googlePublishModel!!)
+                        mqttPublisherViewModel.getDevicesThatConnectedWithMqttPublish(
+                            mqttPublishModel!!.id
+                        )
                             .subscribeOn(Schedulers.io())
                             .observeOn(AndroidSchedulers.mainThread())
                             .subscribe(Consumer {
@@ -244,41 +205,47 @@ class GoogleCloudPublisherActivity : BaseActivity() {
             .isChecked = true
     }
 
-
     private fun setFields() {
-        edit_name.setText(googlePublishModel?.name)
+        edit_name.setText(mqttPublishModel?.name)
 
         edit_interval_count.setText(
             PublisherIntervalConverter.calculateCountFromMillis(
-                googlePublishModel!!.timeMillis,
-                googlePublishModel!!.timeType
+                mqttPublishModel!!.timeMillis,
+                mqttPublishModel!!.timeType
             )
         )
 
         spinner_interval_time.setSelection(
             resources.getStringArray(R.array.PublishIntervals).indexOf(
-                googlePublishModel?.timeType
+                mqttPublishModel?.timeType
             )
         )
 
-        if (googlePublishModel!!.lastTimeMillis == 0L) {
+        if (mqttPublishModel!!.lastTimeMillis == 0L) {
             text_lastdatasent.visibility = View.GONE
         } else {
             text_lastdatasent.visibility = View.VISIBLE
             val str = getString(R.string.last_data_sent) + " " +
                     millisToFormattedDateString(
-                        googlePublishModel!!.lastTimeMillis
+                        mqttPublishModel!!.lastTimeMillis
                     )
             text_lastdatasent.text = str
         }
 
-        edit_projectid.setText(googlePublishModel!!.projectId)
-        edit_region.setText(googlePublishModel!!.region)
-        edit_deviceregistry.setText(googlePublishModel!!.deviceRegistry)
-        edit_device.setText(googlePublishModel!!.device)
-        edit_privatekey.text = googlePublishModel!!.privateKey
 
-        edit_datastring.setText(googlePublishModel?.dataString)
+        edit_url_mqtt.setText(mqttPublishModel!!.url)
+        edit_clientid_mqtt.setText(mqttPublishModel!!.clientId)
+        edit_username_mqtt.setText(mqttPublishModel!!.username)
+        edit_password_mqtt.setText(mqttPublishModel!!.password)
+        edit_topic_mqtt.setText(mqttPublishModel!!.topic)
+
+        when (mqttPublishModel!!.qos) {
+            0 -> qos_0.isChecked = true
+            1 -> qos_1.isChecked = true
+            2 -> qos_2.isChecked = true
+        }
+
+        edit_datastring.setText(mqttPublishModel?.dataString)
     }
 
     private fun millisToFormattedDateString(millis: Long): String {
@@ -289,11 +256,11 @@ class GoogleCloudPublisherActivity : BaseActivity() {
     }
 
     private fun addOrUpdate() {
-        val googlePublishModel = toGooglePublishModel()
-        if (googlePublishModel != null) {
-            googleViewModel.save(googlePublishModel)
+        val mqttPublishModel = toMqttPublishModel()
+        if (mqttPublishModel != null) {
+            mqttPublisherViewModel.save(mqttPublishModel)
                 .flatMapCompletable {
-                    addRelationsToGoogle(it)
+                    addRelationsToMqtt(it)
                 }.subscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread())
                 .subscribe(object : CompletableObserver {
@@ -310,7 +277,7 @@ class GoogleCloudPublisherActivity : BaseActivity() {
                     override fun onError(e: Throwable) {
                         progressbar.visibility = View.INVISIBLE
                         Toast.makeText(
-                            this@GoogleCloudPublisherActivity,
+                            this@MqttPublisherActivity,
                             e.message,
                             Toast.LENGTH_SHORT
                         )
@@ -320,57 +287,10 @@ class GoogleCloudPublisherActivity : BaseActivity() {
         }
     }
 
-    private fun toGooglePublishModel(): GooglePublishModel? {
-        val name = edit_name.text.toString().trim()
-        val projectId = edit_projectid.text.toString().trim()
-        val region = edit_region.text.toString().trim()
-        val deviceRegistry = edit_deviceregistry.text.toString().trim()
-        val device = edit_device.text.toString().trim()
-        val privateKey = edit_privatekey.text.toString().trim()
-        val timeType = spinner_interval_time.selectedItem.toString()
-        val timeCount = edit_interval_count.text.toString()
-        val datastring = edit_datastring.text.toString()
-
-        if (googleViewModel.checkFieldsAreEmpty(
-                name,
-                projectId,
-                region,
-                deviceRegistry,
-                device,
-                privateKey,
-                timeType,
-                timeCount,
-                datastring
-            )
-        ) {
-            Toast.makeText(this, getString(R.string.please_fill_blanks), Toast.LENGTH_SHORT).show()
-            return null
-        }
-
-        val id = if (googlePublishModel == null) 0 else googlePublishModel!!.id
-        val timeMillis = PublisherIntervalConverter.calculateMillis(timeCount, timeType)
-        val lastTimeMillis =
-            if (googlePublishModel == null) 0 else googlePublishModel!!.lastTimeMillis
-        return GooglePublishModel(
-            id,
-            name,
-            projectId,
-            region,
-            deviceRegistry,
-            device,
-            privateKey,
-            false,
-            timeType,
-            timeMillis,
-            lastTimeMillis,
-            datastring
-        )
-    }
-
-    private fun addRelationsToGoogle(gId: Long): Completable {
+    private fun addRelationsToMqtt(it: Long): Completable {
         val count = layout_devices.childCount
 
-        val setOfCompletable: MutableSet<Completable> = mutableSetOf()
+        val setOfCompletable = mutableSetOf<Completable>()
 
         for (i in 0..(count - 1)) {
             val deviceRelationModel = deviceList[i]
@@ -380,65 +300,96 @@ class GoogleCloudPublisherActivity : BaseActivity() {
 
             if (isChecked) {
                 setOfCompletable.add(
-                    googleViewModel.addOrUpdateGoogleRelation(
+                    mqttPublisherViewModel.addOrUpdateMqttRelation(
                         deviceId = deviceRelationModel.macAddress,
-                        googleId = gId
+                        mqttId = it
                     )
                 )
+
             } else {
                 setOfCompletable.add(
-                    googleViewModel.deleteRelationGoogle(
+                    mqttPublisherViewModel.deleteRelationMqtt(
                         deviceId = deviceRelationModel.macAddress,
-                        googleId = gId
+                        mqttId = it
                     )
                 )
             }
         }
+
         return Completable.merge(setOfCompletable)
     }
 
-    private fun isFileValidPKCS8(byteArray: ByteArray): Boolean {
-        val spec = PKCS8EncodedKeySpec(byteArray)
-        val kf = KeyFactory.getInstance("RSA")
+    private fun toMqttPublishModel(): MqttPublishModel? {
+        val name = edit_name.text.toString().trim()
+        val url = edit_url_mqtt.text.toString().trim()
+        val clientId = edit_clientid_mqtt.text.toString().trim()
+        val username = edit_username_mqtt.text.toString().trim()
+        val password = edit_password_mqtt.text.toString().trim()
+        val topic = edit_topic_mqtt.text.toString().trim()
+        val qos =
+            findViewById<RadioButton>(radio_group_mqtt.checkedRadioButtonId).text.toString().toInt()
+        val timeType = spinner_interval_time.selectedItem.toString()
+        val timeCount = edit_interval_count.text.toString()
+        val datastring = edit_datastring.text.toString()
 
-        try {
-            kf.generatePrivate(spec)
-        } catch (e: Exception) {
-            return false
+        if (mqttPublisherViewModel.checkFieldsAreEmpty(
+                name,
+                url,
+                clientId,
+                username,
+                password,
+                topic,
+                timeType,
+                timeCount,
+                datastring
+            )
+        ) {
+            Toast.makeText(this, getString(R.string.please_fill_blanks), Toast.LENGTH_SHORT).show()
+            return null
         }
-        return true
-    }
 
-    private fun getPrivateKeyData(privateKey: String): ByteArray {
-        val uri = Uri.parse(privateKey)
-        val stream = contentResolver.openInputStream(uri)
-
-        val size = stream.available()
-        val buffer = ByteArray(size)
-        stream.read(buffer)
-        stream.close()
-        return buffer
+        val id = if (mqttPublishModel == null) 0 else mqttPublishModel!!.id
+        val timeMillis = PublisherIntervalConverter.calculateMillis(timeCount, timeType)
+        val lastTimeMillis = if (mqttPublishModel == null) 0 else mqttPublishModel!!.lastTimeMillis
+        return MqttPublishModel(
+            id,
+            name,
+            url,
+            clientId,
+            username,
+            password,
+            topic,
+            qos,
+            false,
+            timeType,
+            timeMillis,
+            lastTimeMillis,
+            datastring
+        )
     }
 
     private fun test() {
         if (!isTestingAlreadyRunning) {
             isTestingAlreadyRunning = true
-            Toast.makeText(this, getString(R.string.testings_started), Toast.LENGTH_SHORT).show()
-            val toGooglePublishModel = toGooglePublishModel()
 
-            if (toGooglePublishModel == null) {
+            Toast.makeText(this, getString(R.string.testings_started), Toast.LENGTH_SHORT).show()
+
+            val toMqttPublishModel = toMqttPublishModel()
+
+            if (toMqttPublishModel == null) {
                 isTestingAlreadyRunning = false
                 return
             }
 
-            testGoogleConnection(toGooglePublishModel)
+            testMqttConnection(toMqttPublishModel)
         }
+
     }
 
-    private fun testGoogleConnection(toGooglePublishModel: GooglePublishModel) {
-        val publisher = GoogleCloudPublisher(
+    private fun testMqttConnection(toMqttPublishModel: MqttPublishModel) {
+        val publisher = MqttPublisher(
             applicationContext,
-            GooglePublishModelDataMapper().transform(toGooglePublishModel),
+            MqttPublishModelDataMapper().toMqttPublish(toMqttPublishModel),
             listOf(Device("TestDevice", "Mac"))
         )
 
@@ -446,18 +397,18 @@ class GoogleCloudPublisherActivity : BaseActivity() {
         publisher.test(testConnectionCallback)
     }
 
+
     companion object {
         //This is used for the file selector intent
-        const val PICKFILE_REQUEST_CODE: Int = 10213
-        private const val GOOGLE_PUBLISHER_ACTIVITY_KEY = "GOOGLE_PUBLISHER_ACTIVITY_KEY"
+        private const val MQTT_PUBLISHER_ACTIVITY_KEY = "MQTT_PUBLISHER_ACTIVITY_KEY"
 
-        fun start(context: Context, googlePublishModel: GooglePublishModel? = null) {
-            val intent = Intent(context, GoogleCloudPublisherActivity::class.java)
+        fun start(context: Context, mqttPublishModel: MqttPublishModel? = null) {
+            val intent = Intent(context, MqttPublisherActivity::class.java)
 
-            googlePublishModel?.let {
+            mqttPublishModel?.let {
                 intent.putExtra(
-                    GOOGLE_PUBLISHER_ACTIVITY_KEY,
-                    googlePublishModel
+                    MQTT_PUBLISHER_ACTIVITY_KEY,
+                    mqttPublishModel
                 )
             }
