+++ app/src/main/res/layout/activity_google_cloud_publisher.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="#dfdfdf"
     android:orientation="vertical"
-    tools:context=".ui.settings.selectpublish.GoogleCloudPublisherActivity">
+    tools:context=".ui.settings.publishers.selectpublish.GoogleCloudPublisherActivity">
 
     <android.support.v7.widget.Toolbar
         android:id="@+id/custom_toolbar"
@@ -193,13 +193,13 @@
                         android:layout_marginStart="16dp"
                         android:background="@drawable/bg_rectangle_black"
                         android:gravity="top|start"
-                        android:hint="@string/dataStringHint"
                         android:inputType="textMultiLine"
                         android:lines="8"
                         android:maxLines="10"
                         android:minLines="4"
                         android:padding="5dp"
-                        android:scrollbars="none" />
+                        android:scrollbars="none"
+                        android:text="@string/default_data_string" />
                 </LinearLayout>
             </android.support.v7.widget.CardView>
         </LinearLayout>
