+++ app/src/main/res/layout/activity_publish_list.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ui.settings.PublishListActivity">
+    tools:context=".ui.settings.publishers.PublishListActivity">
 
     <android.support.v7.widget.Toolbar
         android:id="@+id/custom_toolbar"
