+++ app/src/main/res/layout/activity_select_publisher.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ui.settings.selectpublish.SelectPublisherActivity">
+    tools:context=".ui.settings.publishers.selectpublish.SelectPublisherActivity">
 
     <android.support.v7.widget.Toolbar
         android:id="@+id/custom_toolbar"
