+++ app/src/main/res/layout/fragment_add_restheader.xml
@@ -4,7 +4,7 @@
     android:layout_width="wrap_content"
     android:layout_height="wrap_content"
     android:id="@+id/dialog"
-    tools:context=".ui.settings.rheader.AddRESTHeaderDialog">
+    tools:context=".ui.settings.publishers.rheader.AddRESTHeaderDialog">
 
     <LinearLayout
         android:layout_width="300dp"
